Allow use of nolint comments if linter configured (#133)
Some projects prefer using `//nolint` comments instead of configuring
exclusions in the config file. These have the benefit of only disabling
the selected linters at the specific place in the code and not across
the entire file, they are automatically cleaned up when they no longer
apply with `allow-unused: false`, and we can continue to require
explanations at call sites alongside the code failing the lint check.

Update makego to call a bash script `checknolintlint.bash` to verify
that the `nolintlint` linter is enabled and has the standard config. If
this script passes, then allow for `//nolint` comments to be used in the
source code. If not, continue using the existing `//nolint` detection to
fail the build if comments are used.