Rebuild golangci-lint when go version changes
If the project upgrades the version of Go to a new minor (or major)
release but doesn't rebuild golangci-lint with that version of Go, it is
possible that certain language features won't be available until it is
rebuilt.

For example, if golangci-lint is built with Go 1.21, it won't be able to
lint code that uses the new range over integers syntax introduced in
Go 1.22.

Update dep_golangci_lint.mk to add the major and minor version of Go to
the cache key so that it will rebuild golangci-lint automatically if the
Go major or minor version are updated in go.mod.