Run golangci-lint fmt in the format/lint targets
Previously, formatters were run alongside linters when running
`golangci-lint run`. In v2, they are configured and run separately with
the `golangci-lint fmt` command.

Update the `format` target to run the formatter (if golangci-lint is not
disabled). Update the `lint` target to run with `--diff` which will exit
with a non-zero exit code if the code is not formatted.

This should ensure consistency with the behavior prior to upgrading to
v2.